requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anaconda-client==1.11.0
anaconda-navigator==2.3.2
appdirs==1.4.4
async-timeout==4.0.2
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.0
captum==0.6.0
certifi==2022.12.7
clyent==1.2.2
conda==23.1.0
conda-build==3.23.3
conda-repo-cli==1.0.27
conda-verify==3.4.2
constantly==15.1.0
croniter==1.3.8
cssselect==1.1.0
daal4py==2021.5.0
datashape==0.5.4
dateutils==0.6.12
deepdiff==6.2.3
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
fastapi==0.88.0
fonttools==4.25.0
frozenlist==1.3.3
gdown==4.7.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.1
inflection==0.5.1
inquirer==3.1.2
lightning-cloud==0.5.27
llvmlite==0.39.1
markdown-it-py==2.1.0
mccabe==0.6.1
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.3.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openpyxl==3.0.10
ordered-set==4.1.0
orjson==3.8.6
pandas==1.5.2
parso==0.7.0
patsy==0.5.3
pep8==1.7.1
Pillow==9.3.0
protobuf==4.22.1
pyasn1-modules==0.2.8
pycurl==7.45.1
pydantic==1.10.5
PyDispatcher==2.0.5
Pygments==2.14.0
pytest==7.1.2
python-dotenv==0.21.1
python-editor==1.0.4
python-multipart==0.0.5
pytorch-lightning==1.9.2
queuelib==1.5.0
readchar==4.0.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
scikit-learn-intelex==2021.20220215.212715
sip==4.19.13
starlette==0.22.0
starsessions==1.3.0
TBB==0.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
transformers==4.26.1
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
zict==2.1.0



